fix: improve Modal exports
